refactor(core): extract subscription check to shared utility
