+++ apps/api/src/app/v1/workflows/activity-generation/trigger/route.ts
@@ -2,20 +2,9 @@ import { errors } from "@/lib/api-errors";
 import { parseBody } from "@/lib/body-parser";
 import { activityGenerationTriggerSchema } from "@/lib/openapi/schemas/workflows";
 import { activityGenerationWorkflow } from "@/workflows/activity-generation/activity-generation-workflow";
-import { auth } from "@zoonk/core/auth";
-import { findActiveSubscription } from "@zoonk/core/auth/subscription";
+import { hasActiveSubscription } from "@zoonk/core/auth/subscription";
 import { type NextRequest, NextResponse } from "next/server";
 import { start } from "workflow/api";
-
-async function getSubscriptions(request: NextRequest) {
-  try {
-    return await auth.api.listActiveSubscriptions({
-      headers: request.headers,
-    });
-  } catch {
-    return [];
-  }
-}
 
 export async function POST(request: NextRequest) {
   const parsed = await parseBody(request, activityGenerationTriggerSchema);
@@ -24,9 +13,9 @@ export async function POST(request: NextRequest) {
     return errors.validation(parsed.error);
   }
 
-  const subscriptions = await getSubscriptions(request);
+  const hasSubscription = await hasActiveSubscription(request.headers);
 
-  if (!findActiveSubscription(subscriptions)) {
+  if (!hasSubscription) {
     return errors.paymentRequired();
   }
