--- packages/auth/src/hooks/use-subscription.ts
@@ -1,23 +1,0 @@
-import { type Subscription } from "@better-auth/stripe";
-import { useEffect, useState } from "react";
-import { authClient } from "../client";
-import { findActiveSubscription } from "../subscription";
-
-async function getActiveSubscription() {
-  const { data: subscriptions } = await authClient.subscription.list();
-  return findActiveSubscription(subscriptions) ?? null;
-}
-
-export function useSubscription() {
-  const [subscription, setSubscription] = useState<Subscription | null | undefined>();
-
-  const isPending = subscription === undefined;
-
-  useEffect(() => {
-    getActiveSubscription()
-      .then(setSubscription)
-      .catch(() => setSubscription(null));
-  }, []);
-
-  return { isPending, subscription };
-}
